+++ lib/openapi_first/router.rb
@@ -7,20 +7,35 @@ require_relative 'utils'
 module OpenapiFirst
   class Router
     NOT_FOUND = Rack::Response.new('', 404).finish.freeze
+    DEFAULT_NOT_FOUND_APP = ->(_env) { NOT_FOUND }
 
-    def initialize(app, options)
+    def initialize(app, options) # rubocop:disable Metrics/MethodLength
       @app = app
       @namespace = options.fetch(:namespace, nil)
       @parent_app = options.fetch(:parent_app, nil)
-      @router = build_router(options.fetch(:spec).operations)
+      @raise = options.fetch(:raise, false)
+      @failure_app = find_failure_app(options[:not_found])
+      if @failure_app.nil?
+        raise ArgumentError,
+              'not_found must be nil, :continue or must respond to call'
+      end
+      spec = options.fetch(:spec)
+      @filepath = spec.filepath
+      @router = build_router(spec.operations)
     end
 
     def call(env)
       endpoint = find_endpoint(env)
       return endpoint.call(env) if endpoint
+
+      if @raise
+        req = Rack::Request.new(env)
+        msg = "Could not find definition for #{req.request_method} '#{req.path}' in API description #{@filepath}" # rubocop:disable Layout/LineLength
+        raise NotFoundError, msg
+      end
       return @parent_app.call(env) if @parent_app
 
-      NOT_FOUND
+      @failure_app.call(env)
     end
 
     def find_handler(operation_id) # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
@@ -49,6 +64,13 @@ module OpenapiFirst
     end
 
     private
+
+    def find_failure_app(option)
+      return DEFAULT_NOT_FOUND_APP if option.nil?
+      return @app if option == :continue
+
+      option if option.respond_to?(:call)
+    end
 
     def find_endpoint(env)
       original_path_info = env[Rack::PATH_INFO]